Make changes for dart 2

+++ heroes_companion_data/lib/src/models/hero_win_rate.dart
@@ -18,9 +18,10 @@ class HeroWinRate {
   }
 
   factory HeroWinRate.fromWinLossCount(WinLossCount winLossCount, int heroId) {
-    return new HeroWinRate(heroId, winLossCount.winPercentange(),
-        winLossCount.wins + winLossCount.losses);
-  }
+    //TODO change winLossCount.wins to int
+    return new HeroWinRate(heroId, winLossCount.winPercentange().toDouble(),
+        winLossCount.wins.toInt() + winLossCount.losses.toInt());
+  } 
 
   Map toMap() {
     Map map = {
